+++ contract/src/constants.rs
@@ -1,0 +1,3 @@
+pub const TGAS: u64 = 1_000_000_000_000;
+pub const NO_DEPOSIT: u128 = 0;
+pub const NO_ARGS: Vec<u8> = vec![];
